Refactor directory and namespace
