+++ include/urg_stamped/ros_logger.h
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2019 The urg_stamped Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef URG_STAMPED_ROS_LOGGER_H
+#define URG_STAMPED_ROS_LOGGER_H
+
+#include <iostream>
+
+namespace urg_stamped
+{
+void setROSLogger();
+}  // namespace urg_stamped
+
+#endif  // URG_STAMPED_ROS_LOGGER_H
